quotes/requirements.txt
alembic==1.13.1
asgiref==3.7.2
Django==5.0.2
django-environ==0.11.2
greenlet==3.0.3
Mako==1.3.2
MarkupSafe==2.1.5
pillow==10.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.9.0
